pyproject.toml
[build-system]
requires = ["flit_core >=2,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "ellar_django"


[project]
name = "ellar-django-module"
authors = [
   {name = "Ezeudoh Tochukwu", email = "tochukwu.ezeudoh@gmail.com"},
]
dynamic = ["version", "description"]
requires-python = ">=3.8"
readme = "README.md"
home-page = "https://github.com/eadwinCode/ellar-django-module"
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Typing :: Typed",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
    "Framework :: Django",
    "Framework :: Django :: 3.0",
    "Framework :: Django :: 3.1",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: AsyncIO",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "ellar-cli >= 0.1.8",
    "Django >= 3.0",
]


[project.urls]
Documentation = "https://github.com/eadwinCode/ellar-django-module"
Source = "https://github.com/eadwinCode/ellar-django-module"
Homepage = "https://eadwincode.github.io/ellar-django-module/"

[project.optional-dependencies]
test = [
    "pytest >= 7.1.3,<8.0.0",
    "pytest-django",
    "pytest-cov >= 2.12.0,<5.0.0",
    "mypy == 1.4.1",
    "ruff ==0.0.275",
    "black == 22.8.0",
    "pytest-asyncio",
    "autoflake",
]
dev = [
    "pre-commit"
]

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-third-party = ["ellar", "Django"]
